finally resolve #7 - allow saving a nn to a file
With the new save_to_file() and from_file() functions it is now very easy to serialize and deserialize a nn state.
To that end, we introduce a Serializable wrapper to the Network structure.
- Also added a util mod that for now only has one method, which gets the project root directory
- minor tweaks and cleanups
