+++ .gitignore
@@ -5,4 +5,5 @@ Cargo.lock
 logs/*
 *.sublime-*
 data/config.json
+data/*.json
