+++ src/input/util.rs
@@ -1,12 +1,7 @@
 use std::env;
-use std::ffi::OsStr;
-use std::fs;
-use std::io;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
-fn get_root_dir() -> PathBuf {
-    // TODO: try to get the right directory with cargo MANIFEST environment variable
-    let mut p = env::current_dir().unwrap();
-    p.push(Path::new("logs/"));
-    p
+pub fn get_root_dir() -> PathBuf {
+    // TODO: remove expect here
+    PathBuf::from(env::var("CARGO_MANIFEST_DIR").expect("Could not find the CARGO_MANIFEST_DIR"))
 }
